add controller asignatura

+++ src/main/java/com/runt/springboot/backend/models/service/asignatura/IAsignaturaService.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.runt.springboot.backend.asignatura.models.entity.Asignatura;
 
 public interface IAsignaturaService  {
-	public List<Asignatura> findToidTeacher(int id);
+	public List<Asignatura> findToidTeacher(Long id);
 }
